Make aliases on abstract queries required
