+++ packages/data/src/types/abstract-query/common/function/array.ts
@@ -1,0 +1,5 @@
+export interface ArrayFn {
+	type: 'arrayFn';
+
+	fn: 'count';
+}
