+++ packages/data/src/types/abstract-query/common/function/index.ts
@@ -1,0 +1,2 @@
+export * from './array.js';
+export * from './extract.js';
