+++ packages/data/src/types/abstract-query/common/primitive.ts
@@ -1,0 +1,6 @@
+export interface AbstractQueryPrimitive {
+	type: 'primitive';
+
+	/** the name of the attribute */
+	field: string;
+}
