+++ packages/data/src/types/abstract-query/fields.ts
@@ -2,7 +2,6 @@ import type { AbstractQueryFieldNodeFn } from './fields/function.js';
 import type {
 	AbstractQueryFieldNodeNestedMany,
 	AbstractQueryFieldNodeNestedOne,
-	AbstractQueryFieldNodeNestedTarget,
 } from './fields/nested.js';
 import type { AbstractQueryFieldNodePrimitive } from './fields/primitive.js';
 
@@ -11,8 +10,3 @@ export type AbstractQueryFieldNode =
 	| AbstractQueryFieldNodeFn
 	| AbstractQueryFieldNodeNestedMany
 	| AbstractQueryFieldNodeNestedOne;
-
-export type AbstractQueryFieldNodeTarget =
-	| AbstractQueryFieldNodePrimitive
-	| AbstractQueryFieldNodeFn
-	| AbstractQueryFieldNodeNestedTarget;
