+++ packages/data/src/types/abstract-query/fields/index.ts
@@ -1,4 +1,3 @@
 export * from './function.js';
 export * from './nested.js';
-export * from './nested/index.js';
 export * from './primitive.js';
