+++ packages/data/src/types/abstract-query/fields/primitive.ts
@@ -1,5 +1,8 @@
+import type { AbstractQueryPrimitive } from '../common/primitive.js';
+
 /**
  * Generic primitive value read from the store field
+ *
  * @example
  * Let's say you want the engine to only return the `id` field of the collection in question:
  * For that you would create a node like the following and add it to the `nodes` of the query.
@@ -10,11 +13,6 @@
  * }
  * ```
  */
-export interface AbstractQueryFieldNodePrimitive {
-	type: 'primitive';
-
-	/** the name of the attribute */
-	field: string;
-
-	alias?: string;
+export interface AbstractQueryFieldNodePrimitive extends AbstractQueryPrimitive {
+	alias: string;
 }
