+++ packages/data/src/types/abstract-query/index.ts
@@ -1,3 +1,4 @@
+export * from './common/index.js';
 export * from './fields.js';
 export * from './fields/index.js';
 export * from './modifiers.js';
