+++ packages/data/src/types/abstract-query/modifiers/filters/conditions/geo-condition-bbox.ts
@@ -1,12 +1,13 @@
 import type { GeoJSONGeometryCollection, GeoJSONMultiPolygon, GeoJSONPolygon } from 'wellknown';
-import type { AbstractQueryFieldNodeTarget } from '../../../fields.js';
+
+import type { AbstractQueryTarget } from '../../target.js';
 
 /**
  * Used to check if geo box objects intersect
  */
 export interface ConditionGeoIntersectsBBoxNode {
 	type: 'condition-geo-intersects-bbox';
-	target: AbstractQueryFieldNodeTarget;
+	target: AbstractQueryTarget;
 	operation: 'intersects_bbox';
 	compareTo: GeoJSONPolygon | GeoJSONMultiPolygon | GeoJSONGeometryCollection;
 	/** @TODO confirm if MultiPolygon works as expected across drivers */
