+++ packages/data/src/types/abstract-query/modifiers/filters/conditions/geo-condition.ts
@@ -6,7 +6,7 @@ import type {
 	GeoJSONPoint,
 } from 'wellknown';
 
-import type { AbstractQueryFieldNodeTarget } from '../../../fields.js';
+import type { AbstractQueryTarget } from '../../target.js';
 
 /**
  * Checks if a non box geo object intersects with another.
@@ -33,7 +33,7 @@ import type { AbstractQueryFieldNodeTarget } from '../../../fields.js';
  */
 export interface ConditionGeoIntersectsNode {
 	type: 'condition-geo-intersects';
-	target: AbstractQueryFieldNodeTarget /** the type of the field needs to be a 'geometry' object */;
+	target: AbstractQueryTarget /** the type of the field needs to be a 'geometry' object */;
 	operation: 'intersects';
 	compareTo: GeoJSONPoint | GeoJSONMultiPoint | GeoJSONLineString | GeoJSONMultiLineString | GeoJSONGeometryCollection;
 }
