+++ packages/data/src/types/abstract-query/modifiers/filters/conditions/number-condition.ts
@@ -1,4 +1,4 @@
-import type { AbstractQueryFieldNodeTarget } from '../../../fields.js';
+import type { AbstractQueryTarget } from '../../target.js';
 
 /**
  * Used to compare a number or date time field with a number value.
@@ -16,7 +16,7 @@ import type { AbstractQueryFieldNodeTarget } from '../../../fields.js';
  */
 export interface ConditionNumberNode {
 	type: 'condition-number';
-	target: AbstractQueryFieldNodeTarget;
+	target: AbstractQueryTarget;
 	operation: 'eq' | 'lt' | 'lte' | 'gt' | 'gte';
 	compareTo: number;
 }
