+++ packages/data/src/types/abstract-query/modifiers/index.ts
@@ -3,3 +3,5 @@ export * from './filters/index.js';
 export * from './limit.js';
 export * from './offset.js';
 export * from './sort.js';
+export * from './target.js';
+export * from './target/index.js';
